Querying The FIM/MIM Service Using PowerShell
Have you ever needed to query the FIM service for objects based upon an Xpath definition and display the information on screen and if needed also export it to CSV?. Look no further, here’s a PowerShell script just doing that for you!

–

In the blog post (2012-11-11) Finding All Users Within FIM That Have (Not) Registered For SSPR I demonstrate how to query the FIM service using the FIM Portal for users that have registered for SSPR and for users that have not registered for SSPR. This is all done through search scopes. Once I got the question: "can you export that information?". It is by default not possible to export that information through the FIM Portal. However, nothing or nobody stops you from using PowerShell and achieve your goal!

For a detailed explanation of this script, please go to:

(2015-03-30) Querying The FIM Service Using PowerShell