FIM/MIM Configuration Export Scripts
When upgrading FIM or MIM to a newer version, you may need to uninstall the previous version first before installing the newer version. During the installation of the new version you need to reenter all the required information. But where do you get that data from? Either you have some installation/configuration guide or you make sure you make a copy (copy, export, backup) of the previous configuration so that you can look it up easily. This script creates a export of the FIM/MIM.

 

For a detailed explanation of this script, please go to:

(2016-09-25) FIM/MIM Configuration Export Scripts